Add renovate bot to keep kops up to date
This commit adds a configuration for renovate, based on the
configurations found in cilium/little-vm-helper and cilium/cilium, and
modifies the install-kops action to keep the kops version up to date.

Note that the default versioning scheme which will be used is Coerced
SemVar, which does not support pre-releases. Renovate will only update
the kops version when a new major, minor or patch is available.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>