create-cluster: make KCM node CIDR allocations optional
By default, the KCM will automatically allocate a node CIDR for each node
in the cluster. This causes a new static route to be created for each
node when running on GCE, and quickly exceeds the quota limit of 200 per
VPC. Disabling this allows the creation of larger clusters, and since we
are using tunnel mode in the 100 nodes scale test, they are unused
anyway.

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>