Bump kops version to v1.29.0
Although 238d773 ("Add renovate bot to keep kops up to date") enabled
renovate to automatically bump the kops version, it appears that renovate
doesn't automatically bump it if configured to a pre-release version.
Hence, let's manually update it to the latest stable version, so that
subsequent updates can than be performed automatically.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>