create-cluster: increase etcd volume size to 50gb
In a 3k node test, etcd db was filling up. This increases the backend
quota for etcd's database from the 2gb default to 8gb, which is the
maximum recommended size.

Additionally, this increases the size of the volume etcd uses from 20gb
to 50gb. Documentation calls out that IOPS are increased as volume size
increases. We may want to explore increasing further if it seems
prudent.

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>