add action to delete an instance group with retry
When deleting large clusters, kops issues many API requests in a short
period of time. This quickly consumes the quota limit for list and read
requests on GCP and can lead to failures.

Instance groups have a maximum size of 1000 nodes, and large clusters
are deployed using multiple separate instance groups. This action will
help separate the cluster deletion into multiple steps, reducing the
load on the GCP API and lessening the changes of getting rate-limited.

The action is implemented with a retry, using a polling rate of 2
minutes to give the quota time to recover in between retries.

Signed-off-by: Tim Horner <timothy.horner@isovalent.com>