create-cluster/create-instance-group: configure disk type
Configure the disk used for the etcd cluster members to be of type
pd-ssd, to attain better IOPS performance. Similarly, let's make it
possible to configure the disk type of additional instance groups.

According to the comparison provided by the gcloud console, the
following is to be expected for a disk of 50GB (attached to an
instance with 16 cores):

|            | standard  | balanced |  ssd  |
|------------|-----------|----------|-------|
| Read IOPS  |        38 |      300 |  1500 |
| Write IOPS |        75 |      300 | 1,500 |

And similarly for a 200GB disk:

|            | standard  | balanced |  ssd  |
|------------|-----------|----------|-------|
| Read IOPS  |       150 |    1,200 | 6,000 |
| Write IOPS |       300 |    1,200 | 6,000 |

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>