setup-firewall: shorten rule prefix for gcloud constraints
We recently observed the following error when running a scale test:

`Invalid value for field 'resource.name': 'pod-cidrs-to-node-${{ test_name }}-${{ github.run_id }}-${{ github.run_attempt }}-${{ cluster_name }}'. Must be a match of regex '(?:[a-z](?:[-a-z0-9]{0,61}[a-z0-9])?)'`

In particular, the firewall rule name achieved had length=65, which is
more than the allowed by the gcloud sdk (61 characters). Let's shrink the
name of the rules by renaming the prefix from `pod-cidrs-to-node` to
`pods-to-node`. These 5 characters of difference should give us enough
space to run scale tests without this type of error.

Signed-off-by: Simone Magnani <simone.magnani@isovalent.com>